Modify to meet requirements
