+++ prac_08/dynamic_labels.kv
@@ -1,0 +1,5 @@
+BoxLayout:
+    orientation:'horizontal'
+    BoxLayout:
+        orientation: 'vertical'
+        id: main
