--- prac_08/dynamic_widgets.kv
@@ -1,18 +1,0 @@
-BoxLayout:
-    orientation: 'vertical'
-    BoxLayout:
-        orientation: 'vertical'
-        # this layout is populated with Button widgets in app code
-        id: entries_box
-    Label:
-        size_hint_y: 0.1
-        text: app.status_text
-    BoxLayout:
-        orientation: 'horizontal'
-        size_hint_y: 0.2
-        Button:
-            text: "Create Widgets"
-            on_press: app.create_widgets()
-        Button:
-            text: "Clear Widgets"
-            on_press: app.clear_all()
